+++ src/main/java/spleetwaise/commons/ui/PersonListPanel.java
@@ -1,4 +1,4 @@
-package spleetwaise.address.ui;
+package spleetwaise.commons.ui;
 
 import java.util.logging.Logger;
