+++ src/main/java/spleetwaise/transaction/ui/RightPanel.java
@@ -12,8 +12,8 @@ import javafx.scene.control.MenuItem;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.Region;
 import javafx.scene.layout.StackPane;
-import spleetwaise.address.ui.UiPart;
 import spleetwaise.commons.model.CommonModel;
+import spleetwaise.commons.ui.UiPart;
 import spleetwaise.transaction.model.TransactionBookModel;
 import spleetwaise.transaction.model.transaction.Amount;
 import spleetwaise.transaction.model.transaction.Transaction;
