+++ src/test/java/spleetwaise/commons/ui/TestFxmlObject.java
@@ -1,4 +1,4 @@
-package spleetwaise.address.ui;
+package spleetwaise.commons.ui;
 
 import java.util.Objects;
