+++ src/test/java/systemtests/TestFxAppRunner.java
@@ -26,7 +26,7 @@ import spleetwaise.address.model.AddressBook;
 import spleetwaise.address.model.AddressBookModel;
 import spleetwaise.address.testutil.TestUtil;
 import spleetwaise.address.testutil.TypicalPersons;
-import spleetwaise.address.ui.CommandBox;
+import spleetwaise.commons.ui.CommandBox;
 
 /**
  * This class serves as a base class for tests that require the use of the application's GUI and TestFx setup.
