+++ src/test/resources/view/UiPartTest/validFile.fxml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import spleetwaise.address.ui.TestFxmlObject?>
+<?import spleetwaise.commons.ui.TestFxmlObject?>
 <TestFxmlObject xmlns:fx="http://javafx.com/fxml/1" fx:id="validFileRoot">Hello World!</TestFxmlObject>
